Fixed path of components
